database changes

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -17,8 +17,8 @@ class CreateUsersTable extends Migration
         $table->string('first_name');
         $table->string('last_name');
         $table->string('email')->unique();
-        $table->double('longitude', 15, 8);
-        $table->double('latitude', 15, 8);
+        $table->decimal('longitude', 10, 7);
+        $table->decimal('latitude', 10, 7);
         $table->timestamp('email_verified_at')->nullable();
         $table->string('password');
         $table->boolean('is_specialist'); /* 0 for client, 1 for specialist */
